+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Fixed
+-----
+- Fix ``htseq-count-raw`` process to correctly map features with associated
+  feature symbols.
+
+
 =================
 14.0.1 2018-10-23
 =================